KKDF2's Docs simplified
All credit goes to Mr. Kenjiuno for his docs, I have simplified the first parts of it as those parts are concrete fact, while the things closer to the end of his docs get into speculation, though they may be accurate. The files added are the same, I've added the pdf in case you don't want to download the zip