Entire list of KH1 MDLS files listed out with the named characters beside them
The list of files should be credited to Rapid; he compiled the list, I just converted it to markdown format for the repo. Furthermore, I included a couple edits and a "Return to Index" clickable link to every game-specific index document so that users can actually return to the OpenKH.dev homepage instead of having to type it back into the URL bar. Idk why this wasn't a thing before, but here we are now.