Update ConfigurationService.cs
Separate "IsFirstRunComplete" check from base mods-manager.yml file. If changes are made to the mods manager that would require users to go through the setup wizard again we can provide the `first-run-complete.yml` with "isFirstRunComplete=False". This allows us to force the setup wizard to open after a user updates their mod manager. This can come in handy if support for more KH games or more ways to play KH on PC (if steam version comes out one day) is added to the mod manager and we need all users to re-verify their settings. It's also useful for Panacea Mod Loader updates as it can only be installed through the wizard and there's no way to know if there's an update outside of the setup wizard.