Better Detect Install for Steam
If the user had a folder with the same name as where 1.5 2.5 or 2.8 were installed it would say it is a valid install even if the folder is empty. this checks that .dll that lets you press the next button exists in that folder first. EGS detect game install seems to be a little more specific and i do not have the EGS version to test so i did not edit that code.